Split service into interface + default implementation
