+++ saas/node.py
@@ -5,19 +5,22 @@ import time
 from threading import Lock
 from typing import Optional
 
+from saas.dor.api import DORService
+from saas.dor.service import DefaultDORService
 from saas.meta import __version__
 
-import saas.p2p.service as p2p_service
-import saas.dor.service as dor_service
-import saas.rest.service as rest_service
-import saas.rti.service as rti_service
-import saas.nodedb.service as db_service
 from saas.core.helpers import get_timestamp_now
 from saas.core.identity import Identity
 from saas.core.keystore import Keystore
 from saas.core.logging import Logging
+from saas.nodedb.api import NodeDBService
 from saas.nodedb.schemas import NodeInfo
+from saas.nodedb.service import DefaultNodeDBService
 from saas.p2p.exceptions import P2PException, BootNodeUnavailableError
+from saas.p2p.service import P2PService
+from saas.rest.service import RESTService
+from saas.rti.api import RTIService
+from saas.rti.service import DefaultRTIService
 
 logger = Logging.get('node')
 
@@ -30,11 +33,11 @@ class Node:
         self._mutex = Lock()
         self._datastore_path = datastore_path
         self._keystore = keystore
-        self.db: Optional[db_service.NodeDBService] = None
-        self.p2p: Optional[p2p_service.P2PService] = None
-        self.rest: Optional[rest_service.RESTService] = None
-        self.dor: Optional[dor_service.DORService] = None
-        self.rti: Optional[rti_service.RTIService] = None
+        self.p2p: Optional[P2PService] = None
+        self.rest: Optional[RESTService] = None
+        self.db: Optional[NodeDBService] = None
+        self.dor: Optional[DORService] = None
+        self.rti: Optional[RTIService] = None
 
     @property
     def keystore(self) -> Keystore:
@@ -70,34 +73,36 @@ class Node:
         logger.info(f"saas-middleware {__version__}")
 
         logger.info("starting P2P service.")
-        self.p2p = p2p_service.P2PService(self, server_address, bind_all_address)
+        self.p2p = P2PService(self, server_address, bind_all_address)
         self.p2p.start_service()
 
         endpoints = []
         if enable_db:
             db_path = f"sqlite:///{os.path.join(self._datastore_path, 'node.db')}"
             logger.info(f"enabling NodeDB service using {db_path}.")
-            self.db = db_service.NodeDBService(self, db_path)
+            self.db = DefaultNodeDBService(self, db_path)
             self.p2p.add(self.db.protocol)
             endpoints += self.db.endpoints()
 
         if enable_dor:
             db_path = f"sqlite:///{os.path.join(self._datastore_path, 'dor.db')}"
             logger.info(f"enabling DOR service using {db_path}.")
-            self.dor = dor_service.DORService(self, db_path)
+            self.dor = DefaultDORService(self, db_path)
             self.p2p.add(self.dor.protocol)
             endpoints += self.dor.endpoints()
 
         if enable_rti:
             db_path = f"sqlite:///{os.path.join(self._datastore_path, 'rti.db')}"
-            self.rti = rti_service.RTIService(self, db_path, retain_job_history=retain_job_history,
-                                              strict_deployment=strict_deployment, job_concurrency=job_concurrency)
+            self.rti = DefaultRTIService(self, db_path,
+                                         retain_job_history=retain_job_history,
+                                         strict_deployment=strict_deployment,
+                                         job_concurrency=job_concurrency)
             logger.info(f"enabling RTI service using {db_path}.")
             endpoints += self.rti.endpoints()
 
         if rest_address is not None:
             logger.info("starting REST service.")
-            self.rest = rest_service.RESTService(self, rest_address[0], rest_address[1], bind_all_address)
+            self.rest = RESTService(self, rest_address[0], rest_address[1], bind_all_address)
             self.rest.start_service()
             self.rest.add(endpoints)
