+++ tests/test_cli.py
@@ -27,11 +27,11 @@ from saas.cli.exceptions import CLIRuntimeError
 from saas.core.identity import Identity
 from saas.core.keystore import Keystore
 from saas.core.logging import Logging
-from saas.dor.proxy import DORProxy
+from saas.dor.api import DORProxy
 from saas.dor.schemas import DataObject, ProcessorDescriptor, GitProcessorPointer
 from saas.helpers import find_available_port, docker_export_image
 from saas.node import Node
-from saas.rti.proxy import JobRESTProxy
+from saas.rti.api import JobRESTProxy
 from saas.rti.schemas import Task, Job, JobStatus, Severity, ExitCode, JobResult, Processor
 from saas.core.processor import ProgressListener, ProcessorBase, ProcessorRuntimeError, find_processors
 from tests.base_testcase import PortMaster
